requirements.txt
langchain
langchain-openai
langchain-community
faiss-cpu
openai
beautifulsoup4
python-dotenv
requests
